+++ Docs/CHANGELOG.md
@@ -24,6 +24,7 @@
     - VehicleOpensDoor: a parked vehicle next to the ego suddenly opens the door, forcing the ego to brake. The ego has to maneuver to an adjacent lane to surpass the obstacle.
     - HighwayCutIn: The ego is met with a vehicle that tries to enter the highway by cutting in front of it.
     - ParkingCutIn: A vehicle parked at the side cuts in front of the ego.
+    - MergerIntoSlowTraffic: the ego has to enter a lane that is filled with slow traffic.
 * Added new functions to the BackgroundManager
 * Minor improvements to some example scenarios. These include FollowLeadingVehicle, VehicleTurning, DynamicObjectCrossing and SignalizedJunctionRightTurn and RunningRedLight. Their behaviors are now more smooth, robust and some outdated mechanics have been removed
 * SignalizedJunctionLeftTurn has been remade. It now has an actor flow on which the ego has to merge into, instead of a single vehicle.